@@ -50,79 +68,75 @@
     "prop-types": "^15.7.2",
     "qs": "^6.7.0",
     "queue": "^6.0.1",
-    "rc-animate": "^2.8.3",
-    "react": "^16.8.6",
+    "react": "^16.9.0",
     "react-amap": "^1.2.8",
-    "react-amap-plugin-geolocation": "^0.0.2",
     "react-container-query": "^0.11.0",
     "react-copy-to-clipboard": "^5.0.1",
     "react-document-title": "^2.0.3",
-    "react-dom": "^16.8.6",
+    "react-dom": "^16.9.0",
     "react-fittext": "^1.0.0",
     "react-infinite-scroller": "^1.2.4",
     "react-media": "^1.9.2",
+    "react-media-hook2": "^1.1.2",
     "react-router-dom": "^5.0.1",
-    "react-zmage": "^0.8.1",
-    "string": "^3.3.3",
+    "react-zmage": "^0.8.4",
+    "redux": "^4.0.4",
+    "umi": "^2.8.18",
+    "umi-plugin-pro-block": "^1.3.4",
+    "umi-plugin-react": "^1.9.17",
+    "umi-request": "^1.2.2",
     "uuid": "^3.3.2"
   },
   "devDependencies": {
-    "@types/react": "^16.8.22",
-    "@types/react-dom": "^16.8.4",
-    "antd-pro-merge-less": "^1.0.0",
-    "antd-theme-webpack-plugin": "^1.3.0",
-    "babel-eslint": "^10.0.2",
+    "@ant-design/colors": "^3.1.0",
+    "@ant-design/pro-cli": "^1.0.10",
+    "@types/classnames": "^2.2.9",
+    "@types/express": "^4.17.0",
+    "@types/history": "^4.7.2",
+    "@types/jest": "^24.0.17",
+    "@types/lodash": "^4.14.136",
+    "@types/qs": "^6.5.3",
+    "@types/react": "^16.9.1",
+    "@types/react-document-title": "^2.0.3",
+    "@types/react-dom": "^16.8.5",
+    "@umijs/fabric": "^1.1.9",
+    "chalk": "^2.4.2",
     "cross-env": "^5.2.0",
     "cross-port-killer": "^1.1.1",
     "enzyme": "^3.10.0",
-    "eslint": "^6.0.1",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-config-prettier": "^6.0.0",
-    "eslint-plugin-babel": "^5.3.0",
-    "eslint-plugin-compat": "^3.2.0",
-    "eslint-plugin-import": "^2.18.0",
-    "eslint-plugin-jsx-a11y": "^6.2.1",
-    "eslint-plugin-markdown": "^1.0.0",
-    "eslint-plugin-react": "^7.14.2",
-    "tslint": "^5.18.0",
-    "tslint-config-prettier": "^1.18.0",
-    "tslint-react": "^4.0.0",
-    "gh-pages": "^2.0.1",
-    "husky": "^2.7.0",
-    "lint-staged": "^8.2.1",
-    "merge-umi-mock-data": "^2.0.6",
+    "eslint": "^6.1.0",
+    "gh-pages": "^2.1.1",
+    "husky": "^3.0.3",
+    "import-sort-cli": "^6.0.0",
+    "import-sort-parser-babylon": "^6.0.0",
+    "import-sort-parser-typescript": "^6.0.0",
+    "import-sort-style-module": "^6.0.0",
+    "jest-puppeteer": "^4.3.0",
+    "lint-staged": "^9.2.1",
     "mockjs": "^1.0.1-beta3",
-    "prettier": "1.18.2",
-    "pro-download": "^1.0.1",
+    "node-fetch": "^2.6.0",
+    "prettier": "^1.18.2",
+    "pro-download": "1.0.1",
+    "slash2": "^2.0.0",
     "stylelint": "^10.1.0",
     "stylelint-config-prettier": "^5.2.0",
     "stylelint-config-standard": "^18.3.0",
-    "umi": "^2.8.7",
-    "umi-plugin-ga": "^1.1.3",
-    "umi-plugin-react": "^1.9.6"
+    "umi-plugin-ga": "^1.1.5",
+    "umi-plugin-pro": "^1.0.2",
+    "umi-types": "^0.3.17",
+    "webpack-theme-color-replacer": "^1.2.17"
   },
   "optionalDependencies": {
-    "puppeteer": "^1.18.1"
-  },
-  "lint-staged": {
-    "**/*.{js,jsx,less}": [
-      "prettier --write",
-      "git add"
-    ],
-    "**/*.{js,jsx}": "npm run lint-staged:js",
-    "**/*.less": "stylelint --syntax less"
+    "puppeteer": "^1.19.0"
   },
   "engines": {
-    "node": ">=8.0.0"
+    "node": ">=10.0.0"
   },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not ie <= 10"
-  ],
-  "husky": {
-    "hooks": {
-      "pre-commit": "npm run lint-staged"
-    }
-  }
+  "checkFiles": [
+    "src/**/*.js*",
+    "src/**/*.ts*",
+    "src/**/*.less",
+    "config/**/*.js*",
+    "scripts/**/*.js"
+  ]
 }
