@@ -6,5 +6,83 @@ module.exports = {
   fixedHeader: false, // sticky header
   autoHideHeader: false, // auto hide header
   fixSiderbar: false, // sticky siderbar
-  title: '量子美食 商城管理系统',
+  defaultEmptyImage: '/noImageSmall.png', // sticky siderbar
+  getPlatformName: () => {
+    let result = '';
+
+    if ((window.appInitCustom || null) != null) {
+      if ((window.appInitCustom.appName || null) != null) {
+        const { platformName } = window.appInitCustom;
+
+        result = platformName;
+      }
+    }
+
+    return result || '';
+  },
+  getTitle: () => {
+    let result = '商城管理系统';
+
+    if ((window.appInitCustom || null) != null) {
+      if ((window.appInitCustom.appName || null) != null) {
+        const { appName } = window.appInitCustom;
+
+        result = appName;
+      }
+    }
+
+    return result || '';
+  },
+  getLoginLogo: () => {
+    let result = '/Logo.png';
+
+    if ((window.appInitCustom || null) != null) {
+      if ((window.appInitCustom.appName || null) != null) {
+        const { loginLogo } = window.appInitCustom;
+
+        result = loginLogo;
+      }
+    }
+
+    return result || '';
+  },
+  getShareLogo: () => {
+    let result = '/shareLogo.png';
+
+    if ((window.appInitCustom || null) != null) {
+      if ((window.appInitCustom.appName || null) != null) {
+        const { shareLogo } = window.appInitCustom;
+
+        result = shareLogo;
+      }
+    }
+
+    return result || '';
+  },
+  getShareLogoName: () => {
+    let result = '';
+
+    if ((window.appInitCustom || null) != null) {
+      if ((window.appInitCustom.appName || null) != null) {
+        const { shareLogoName } = window.appInitCustom;
+
+        result = shareLogoName;
+      }
+    }
+
+    return result || '';
+  },
+  getCompanyName: () => {
+    let result = '';
+
+    if ((window.appInitCustom || null) != null) {
+      if ((window.appInitCustom.appName || null) != null) {
+        const { companyName } = window.appInitCustom;
+
+        result = companyName;
+      }
+    }
+
+    return result || '';
+  },
 };
