@@ -1,293 +1,20 @@
+import component from './en-US/component';
+import globalHeader from './en-US/globalHeader';
+import menu from './en-US/menu';
+import pwa from './en-US/pwa';
+import settingDrawer from './en-US/settingDrawer';
+import settings from './en-US/settings';
+
 export default {
   'navBar.lang': 'Languages',
   'layout.user.link.help': 'Help',
   'layout.user.link.privacy': 'Privacy',
   'layout.user.link.terms': 'Terms',
-  'validation.email.required': 'Please enter your email!',
-  'validation.email.wrong-format': 'The email address is in the wrong format!',
-  'validation.password.required': 'Please enter your password!',
-  'validation.password.twice': 'The passwords entered twice do not match!',
-  'validation.password.strength.msg':
-    "Please enter at least 6 characters and don't use passwords that are easy to guess.",
-  'validation.password.strength.strong': 'Strength: strong',
-  'validation.password.strength.medium': 'Strength: medium',
-  'validation.password.strength.short': 'Strength: too short',
-  'validation.confirm-password.required': 'Please confirm your password!',
-  'validation.phone-number.required': 'Please enter your phone number!',
-  'validation.phone-number.wrong-format': 'Malformed phone number!',
-  'validation.verification-code.required': 'Please enter the verification code!',
-  'validation.title.required': 'Please enter a title',
-  'validation.date.required': 'Please select the start and end date',
-  'validation.goal.required': 'Please enter a description of the goal',
-  'validation.standard.required': 'Please enter a metric',
-  'form.optional': ' (optional) ',
-  'form.submit': 'Submit',
-  'form.save': 'Save',
-  'form.email.placeholder': 'Email',
-  'form.password.placeholder': 'Password',
-  'form.confirm-password.placeholder': 'Confirm password',
-  'form.phone-number.placeholder': 'Phone number',
-  'form.verification-code.placeholder': 'Verification code',
-  'form.title.label': 'Title',
-  'form.title.placeholder': 'Give the target a name',
-  'form.date.label': 'Start and end date',
-  'form.date.placeholder.start': 'Start date',
-  'form.date.placeholder.end': 'End date',
-  'form.goal.label': 'Goal description',
-  'form.goal.placeholder': 'Please enter your work goals',
-  'form.standard.label': 'Metrics',
-  'form.standard.placeholder': 'Please enter a metric',
-  'form.client.label': 'Client',
-  'form.client.label.tooltip': 'Target service object',
-  'form.client.placeholder':
-    'Please describe your customer service, internal customers directly @ Name / job number',
-  'form.invites.label': 'Inviting critics',
-  'form.invites.placeholder': 'Please direct @ Name / job number, you can invite up to 5 people',
-  'form.weight.label': 'Weight',
-  'form.weight.placeholder': 'Please enter weight',
-  'form.public.label': 'Target disclosure',
-  'form.public.label.help': 'Customers and invitees are shared by default',
-  'form.public.radio.public': 'Public',
-  'form.public.radio.partially-public': 'Partially public',
-  'form.public.radio.private': 'Private',
-  'form.publicUsers.placeholder': 'Open to',
-  'form.publicUsers.option.A': 'Colleague A',
-  'form.publicUsers.option.B': 'Colleague B',
-  'form.publicUsers.option.C': 'Colleague C',
-  'component.globalHeader.search': 'Search',
-  'component.globalHeader.search.example1': 'Search example 1',
-  'component.globalHeader.search.example2': 'Search example 2',
-  'component.globalHeader.search.example3': 'Search example 3',
-  'component.globalHeader.help': 'Help',
-  'component.globalHeader.notification': 'Notification',
-  'component.globalHeader.notification.empty': 'You have viewed all notifications.',
-  'component.globalHeader.message': 'Message',
-  'component.globalHeader.message.empty': 'You have viewed all messages.',
-  'component.globalHeader.event': 'Event',
-  'component.globalHeader.event.empty': 'You have viewed all events.',
-  'component.noticeIcon.clear': 'Clear',
-  'component.noticeIcon.cleared': 'Cleared',
-  'component.noticeIcon.empty': 'No notifications',
-  'menu.home': 'Home',
-  'menu.dashboard': 'Dashboard',
-  'menu.dashboard.analysis': 'Analysis',
-  'menu.dashboard.monitor': 'Monitor',
-  'menu.dashboard.workplace': 'Workplace',
-  'menu.form': 'Form',
-  'menu.form.basicForm': 'Basic Form',
-  'menu.form.stepForm': 'Step Form',
-  'menu.form.stepForm.info': 'Step Form(write transfer information)',
-  'menu.form.stepForm.confirm': 'Step Form(confirm transfer information)',
-  'menu.form.stepForm.result': 'Step Form(finished)',
-  'menu.form.advancedForm': 'Advanced Form',
-  'menu.list': 'List',
-  'menu.list.searchTable': 'Search Table',
-  'menu.list.basicList': 'Basic List',
-  'menu.list.cardList': 'Card List',
-  'menu.list.searchList': 'Search List',
-  'menu.list.searchList.articles': 'Search List(articles)',
-  'menu.list.searchList.projects': 'Search List(projects)',
-  'menu.list.searchList.applications': 'Search List(applications)',
-  'menu.profile': 'Profile',
-  'menu.profile.basic': 'Basic Profile',
-  'menu.profile.advanced': 'Advanced Profile',
-  'menu.result': 'Result',
-  'menu.result.success': 'Success',
-  'menu.result.fail': 'Fail',
-  'menu.exception': 'Exception',
-  'menu.exception.not-permission': '403',
-  'menu.exception.not-find': '404',
-  'menu.exception.server-error': '500',
-  'menu.exception.trigger': 'Trigger',
-  'menu.account': 'Account',
-  'menu.account.center': 'Account Center',
-  'menu.account.settings': 'Account Settings',
-  'menu.account.trigger': 'Trigger Error',
-  'menu.account.logout': 'Logout',
-  'app.login.message-invalid-credentials': 'Invalid username or password（admin/888888）',
-  'app.login.message-invalid-verification-code': 'Invalid verification code',
-  'app.login.tab-login-credentials': 'Credentials',
-  'app.login.tab-login-mobile': 'Mobile number',
-  'app.login.remember-me': 'Remember me',
-  'app.login.forgot-password': 'Forgot your password?',
-  'app.login.sign-in-with': 'Sign in with',
-  'app.login.signup': 'Sign up',
-  'app.login.login': 'Login',
-  'app.register.register': 'Register',
-  'app.register.get-verification-code': 'Get code',
-  'app.register.sing-in': 'Already have an account?',
-  'app.register-result.msg': 'Account：registered at {email}',
-  'app.register-result.activation-email':
-    'The activation email has been sent to your email address and is valid for 24 hours. Please log in to the email in time and click on the link in the email to activate the account.',
-  'app.register-result.back-home': 'Back to home',
-  'app.register-result.view-mailbox': 'View mailbox',
-  'app.home.introduce': 'introduce',
-  'app.analysis.test': 'Gongzhuan No.{no} shop',
-  'app.analysis.introduce': 'Introduce',
-  'app.analysis.total-sales': 'Total Sales',
-  'app.analysis.day-sales': 'Day Sales',
-  'app.analysis.visits': 'Visits',
-  'app.analysis.visits-trend': 'Visits Trend',
-  'app.analysis.visits-ranking': 'Visits Ranking',
-  'app.analysis.day-visits': 'Day Visits',
-  'app.analysis.week': 'Week Ratio',
-  'app.analysis.day': 'Day Ratio',
-  'app.analysis.payments': 'Payments',
-  'app.analysis.conversion-rate': 'Conversion Rate',
-  'app.analysis.operational-effect': 'Operational Effect',
-  'app.analysis.sales-trend': 'Stores Sales Trend',
-  'app.analysis.sales-ranking': 'Sales Ranking',
-  'app.analysis.all-year': 'All Year',
-  'app.analysis.all-month': 'All Month',
-  'app.analysis.all-week': 'All Week',
-  'app.analysis.all-day': 'All day',
-  'app.analysis.search-users': 'Search Users',
-  'app.analysis.per-capita-search': 'Per Capita Search',
-  'app.analysis.online-top-search': 'Online Top Search',
-  'app.analysis.the-proportion-of-sales': 'The Proportion Of Sales',
-  'app.analysis.channel.all': 'ALL',
-  'app.analysis.channel.online': 'Online',
-  'app.analysis.channel.stores': 'Stores',
-  'app.analysis.sales': 'Sales',
-  'app.analysis.traffic': 'Traffic',
-  'app.analysis.table.rank': 'Rank',
-  'app.analysis.table.search-keyword': 'Keyword',
-  'app.analysis.table.users': 'Users',
-  'app.analysis.table.weekly-range': 'Weekly Range',
-  'app.forms.basic.title': 'Basic form',
-  'app.forms.basic.description':
-    'Form pages are used to collect or verify information to users, and basic forms are common in scenarios where there are fewer data items.',
-  'app.monitor.trading-activity': 'Real-Time Trading Activity',
-  'app.monitor.total-transactions': 'Total transactions today',
-  'app.monitor.sales-target': 'Sales target completion rate',
-  'app.monitor.remaining-time': 'Remaining time of activity',
-  'app.monitor.total-transactions-per-second': 'Total transactions per second',
-  'app.monitor.activity-forecast': 'Activity forecast',
-  'app.monitor.efficiency': 'Efficiency',
-  'app.monitor.ratio': 'Ratio',
-  'app.monitor.proportion-per-category': 'Proportion Per Category',
-  'app.monitor.fast-food': 'Fast food',
-  'app.monitor.western-food': 'Western food',
-  'app.monitor.hot-pot': 'Hot pot',
-  'app.monitor.waiting-for-implementation': 'Waiting for implementation',
-  'app.monitor.popular-searches': 'Popular Searches',
-  'app.monitor.resource-surplus': 'Resource Surplus',
-  'app.monitor.fund-surplus': 'Fund Surplus',
-  'app.settings.menuMap.basic': 'Basic Settings',
-  'app.settings.menuMap.security': 'Security Settings',
-  'app.settings.menuMap.binding': 'Account Binding',
-  'app.settings.menuMap.notification': 'New Message Notification',
-  'app.settings.basic.avatar': 'Avatar',
-  'app.settings.basic.change-avatar': 'Change avatar',
-  'app.settings.basic.email': 'Email',
-  'app.settings.basic.email-message': 'Please input your email!',
-  'app.settings.basic.nickname': 'Nickname',
-  'app.settings.basic.nickname-message': 'Please input your Nickname!',
-  'app.settings.basic.profile': 'Personal profile',
-  'app.settings.basic.profile-message': 'Please input your personal profile!',
-  'app.settings.basic.profile-placeholder': 'Brief introduction to yourself',
-  'app.settings.basic.country': 'Country/Region',
-  'app.settings.basic.country-message': 'Please input your country!',
-  'app.settings.basic.geographic': 'Province or city',
-  'app.settings.basic.geographic-message': 'Please input your geographic info!',
-  'app.settings.basic.address': 'Street Address',
-  'app.settings.basic.address-message': 'Please input your address!',
-  'app.settings.basic.phone': 'Phone Number',
-  'app.settings.basic.phone-message': 'Please input your phone!',
-  'app.settings.basic.update': 'Update Information',
-  'app.settings.security.strong': 'Strong',
-  'app.settings.security.medium': 'Medium',
-  'app.settings.security.weak': 'Weak',
-  'app.settings.security.password': 'Account Password',
-  'app.settings.security.password-description': 'Current password strength：',
-  'app.settings.security.phone': 'Security Phone',
-  'app.settings.security.phone-description': 'Bound phone：',
-  'app.settings.security.question': 'Security Question',
-  'app.settings.security.question-description':
-    'The security question is not set, and the security policy can effectively protect the account security',
-  'app.settings.security.email': 'Backup Email',
-  'app.settings.security.email-description': 'Bound Email：',
-  'app.settings.security.mfa': 'MFA Device',
-  'app.settings.security.mfa-description':
-    'Unbound MFA device, after binding, can be confirmed twice',
-  'app.settings.security.modify': 'Modify',
-  'app.settings.security.set': 'Set',
-  'app.settings.security.bind': 'Bind',
-  'app.settings.binding.taobao': 'Binding Taobao',
-  'app.settings.binding.taobao-description': 'Currently unbound Taobao account',
-  'app.settings.binding.alipay': 'Binding Alipay',
-  'app.settings.binding.alipay-description': 'Currently unbound Alipay account',
-  'app.settings.binding.dingding': 'Binding DingTalk',
-  'app.settings.binding.dingding-description': 'Currently unbound DingTalk account',
-  'app.settings.binding.bind': 'Bind',
-  'app.settings.notification.password': 'Account Password',
-  'app.settings.notification.password-description':
-    'Messages from other users will be notified in the form of a station letter',
-  'app.settings.notification.messages': 'System Messages',
-  'app.settings.notification.messages-description':
-    'System messages will be notified in the form of a station letter',
-  'app.settings.notification.todo': 'To-do Notification',
-  'app.settings.notification.todo-description':
-    'The to-do list will be notified in the form of a letter from the station',
-  'app.settings.open': 'Open',
-  'app.settings.close': 'Close',
-  'app.exception.back': 'Back to home',
-  'app.exception.description.403': "Sorry, you don't have access to this page",
-  'app.exception.description.404': 'Sorry, the page you visited does not exist',
-  'app.exception.description.500': 'Sorry, the server is reporting an error',
-  'app.result.error.title': 'Submission Failed',
-  'app.result.error.description':
-    'Please check and modify the following information before resubmitting.',
-  'app.result.error.hint-title': 'The content you submitted has the following error:',
-  'app.result.error.hint-text1': 'Your account has been frozen',
-  'app.result.error.hint-btn1': 'Thaw immediately',
-  'app.result.error.hint-text2': 'Your account is not yet eligible to apply',
-  'app.result.error.hint-btn2': 'Upgrade immediately',
-  'app.result.error.btn-text': 'Return to modify',
-  'app.result.success.title': 'Submission Success',
-  'app.result.success.description':
-    'The submission results page is used to feed back the results of a series of operational tasks. If it is a simple operation, use the Message global prompt feedback. This text area can show a simple supplementary explanation. If there is a similar requirement for displaying “documents”, the following gray area can present more complicated content.',
-  'app.result.success.operate-title': 'Project Name',
-  'app.result.success.operate-id': 'Project ID：',
-  'app.result.success.principal': 'Principal：',
-  'app.result.success.operate-time': 'Effective time：',
-  'app.result.success.step1-title': 'Create project',
-  'app.result.success.step1-operator': 'Qu Lili',
-  'app.result.success.step2-title': 'Departmental preliminary review',
-  'app.result.success.step2-operator': 'Zhou Maomao',
-  'app.result.success.step2-extra': 'Urge',
-  'app.result.success.step3-title': 'Financial review',
-  'app.result.success.step4-title': 'Finish',
-  'app.result.success.btn-return': 'Back to list',
-  'app.result.success.btn-project': 'View project',
-  'app.result.success.btn-print': 'Print',
-  'app.setting.pagestyle': 'Page style setting',
-  'app.setting.pagestyle.dark': 'Dark style',
-  'app.setting.pagestyle.light': 'Light style',
-  'app.setting.content-width': 'Content Width',
-  'app.setting.content-width.fixed': 'Fixed',
-  'app.setting.content-width.fluid': 'Fluid',
-  'app.setting.themecolor': 'Theme Color',
-  'app.setting.themecolor.dust': 'Dust Red',
-  'app.setting.themecolor.volcano': 'Volcano',
-  'app.setting.themecolor.sunset': 'Sunset Orange',
-  'app.setting.themecolor.cyan': 'Cyan',
-  'app.setting.themecolor.green': 'Polar Green',
-  'app.setting.themecolor.daybreak': 'Daybreak Blue (default)',
-  'app.setting.themecolor.geekblue': 'Geek Glue',
-  'app.setting.themecolor.purple': 'Golden Purple',
-  'app.setting.navigationmode': 'Navigation Mode',
-  'app.setting.sidemenu': 'Side Menu Layout',
-  'app.setting.topmenu': 'Top Menu Layout',
-  'app.setting.fixedheader': 'Fixed Header',
-  'app.setting.fixedsidebar': 'Fixed Sidebar',
-  'app.setting.fixedsidebar.hint': 'Works on Side Menu Layout',
-  'app.setting.hideheader': 'Hidden Header when scrolling',
-  'app.setting.hideheader.hint': 'Works when Hidden Header is enabled',
-  'app.setting.othersettings': 'Other Settings',
-  'app.setting.weakmode': 'Weak Mode',
-  'app.setting.copy': 'Copy Setting',
-  'app.setting.copyinfo': 'copy success，please replace defaultSettings in src/models/setting.js',
-  'app.setting.production.hint':
-    'Setting panel shows in development environment only, please manually modify',
+  'app.preview.down.block': 'Download this page to your local project',
+  ...globalHeader,
+  ...menu,
+  ...settingDrawer,
+  ...settings,
+  ...pwa,
+  ...component,
 };
