@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import { connect } from 'dva';
 import router from 'umi/router';
-import { FormattedMessage } from 'umi/locale';
+import { FormattedMessage } from 'umi-plugin-react/locale';
 import { Menu } from 'antd';
-import GridContent from '@/components/PageHeaderWrapper/GridContent';
+import { GridContent } from '@ant-design/pro-layout';
 import styles from './index.less';
 
 const { Item } = Menu;
