@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'dva';
 import router from 'umi/router';
 import { Menu } from 'antd';
-import GridContent from '@/components/PageHeaderWrapper/GridContent';
+import { GridContent } from '@ant-design/pro-layout';
 import styles from './index.less';
 
 const { Item } = Menu;
