@@ -4,9 +4,9 @@ import { connect } from 'dva';
 import Link from 'umi/link';
 import { Row, Col, Card, List, Avatar } from 'antd';
 
-import { Radar } from '@/components/Charts';
+import { Radar } from '@/customComponents/Charts';
 import EditableLinkGroup from '@/components/EditableLinkGroup';
-import PageHeaderWrapper from '@/components/PageHeaderWrapper';
+import PageHeaderWrapper from '@/customComponents/PageHeaderWrapper';
 
 import styles from './Workplace.less';
 
