@@ -1,5 +1,7 @@
-/* globals workbox */
+/* eslint-disable eslint-comments/disable-enable-pair */
 /* eslint-disable no-restricted-globals */
+/* eslint-disable no-underscore-dangle */
+/* globals workbox */
 workbox.core.setCacheNameDetails({
   prefix: 'antd-pro',
   suffix: 'v1',
@@ -11,7 +13,6 @@ workbox.clientsClaim();
  * Use precaching list generated by workbox in build process.
  * https://developers.google.com/web/tools/workbox/reference-docs/latest/workbox.precaching
  */
-/* eslint-disable no-underscore-dangle */
 workbox.precaching.precacheAndRoute(self.__precacheManifest || []);
 
 /**
@@ -38,11 +39,11 @@ workbox.routing.registerRoute(/\/api\//, workbox.strategies.networkFirst());
  */
 workbox.routing.registerRoute(
   /^https:\/\/gw.alipayobjects.com\//,
-  workbox.strategies.networkFirst()
+  workbox.strategies.networkFirst(),
 );
 workbox.routing.registerRoute(
   /^https:\/\/cdnjs.cloudflare.com\//,
-  workbox.strategies.networkFirst()
+  workbox.strategies.networkFirst(),
 );
 workbox.routing.registerRoute(/\/color.less/, workbox.strategies.networkFirst());
 
@@ -58,8 +59,8 @@ addEventListener('message', event => {
         .skipWaiting()
         .then(
           () => replyPort.postMessage({ error: null }),
-          error => replyPort.postMessage({ error })
-        )
+          error => replyPort.postMessage({ error }),
+        ),
     );
   }
 });
