@@ -1,11 +1,13 @@
-import RenderAuthorized from '@/components/Authorized';
+import RenderAuthorize from '@/components/Authorized';
 import { getAuthority } from './authority';
+/* eslint-disable eslint-comments/disable-enable-pair */
 
-let Authorized = RenderAuthorized(getAuthority()); // eslint-disable-line
+/* eslint-disable import/no-mutable-exports */
 
-// Reload the rights component
+let Authorized = RenderAuthorize(getAuthority()); // Reload the rights component
+
 const reloadAuthorized = () => {
-  Authorized = RenderAuthorized(getAuthority());
+  Authorized = RenderAuthorize(getAuthority());
 };
 
 export { reloadAuthorized };
