@@ -1,3 +1,6 @@
+/* eslint-disable eslint-comments/disable-enable-pair */
+/* eslint-disable @typescript-eslint/no-var-requires */
+/* eslint-disable eslint-comments/no-unlimited-disable */
 const { spawn } = require('child_process');
 const { kill } = require('cross-port-killer');
 
@@ -20,18 +23,27 @@ startServer.on('exit', () => {
   kill(process.env.PORT || 8000);
 });
 
-// eslint-disable-next-line
+// eslint-disable-next-line no-console
 console.log('Starting development server for e2e tests...');
 startServer.stdout.on('data', data => {
-  // eslint-disable-next-line
+  // eslint-disable-next-line no-console
   console.log(data.toString());
-  if (!once && data.toString().indexOf('App running at') >= 0) {
+  // hack code , wait umi
+  if (
+    (!once && data.toString().indexOf('Compiled successfully') >= 0) ||
+    data.toString().indexOf('Theme generated successfully') >= 0
+  ) {
     // eslint-disable-next-line
     once = true;
+    // eslint-disable-next-line no-console
     console.log('Development server is started, ready to run tests.');
-    const testCmd = spawn(/^win/.test(process.platform) ? 'npm.cmd' : 'npm', ['test'], {
-      stdio: 'inherit',
-    });
+    const testCmd = spawn(
+      /^win/.test(process.platform) ? 'npm.cmd' : 'npm',
+      ['test', '--', '--maxWorkers=1', '--runInBand'],
+      {
+        stdio: 'inherit',
+      },
+    );
     testCmd.on('exit', code => {
       startServer.kill();
       process.exit(code);
